.claude/skills/meow:web-to-markdown/scripts/requirements.txt
# meow:web-to-markdown — Python dependencies
# Install into skill venv: .claude/skills/.venv/bin/pip install -r requirements.txt
# No Playwright here — installed separately via mewkit CLI (system-deps-registry)

# Core HTTP client — used by _static_fetch, fetch_api_spec
requests>=2.31.0

# Readability-based main-content extraction — used by _extract_main_content
readability-lxml>=0.8.1

# HTML → markdown conversion — used by _html_to_markdown
html2text>=2020.1.16

# XML/HTML parsing with injection-safe huge_tree=False — used by _candidate_html_blocks
lxml>=5.1.0

# Charset detection (used by requests for apparent_encoding) — explicit pin for determinism
charset-normalizer>=3.3.0

# YAML parsing — used by injection_detect.py to load vendored
# pyyaml is a soft dependency; injection_detect.py falls back to hardcoded patterns if unavailable
PyYAML>=6.0.1
